Parameterized platform and platform-shard
Previously, the shard was extracted out of the player data map.
Because of the API's current distinction between platform and
platform-shard, a new approach is required to allow the client to
request older season data.

This probably isn't *useful* generally because the platform-shard was
only required for legacy season data. The need for this is very, very
narrow.